pyproject.toml
[tool.poetry]
name = "deepfake-detection"
version = "0.1.0"
description = ""
authors = ["Faraz Ghasemzadeh <faraz.g2000@gmail.com>"]
packages = [{include = "deepfake_detection"}]

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.24.0"
pandas = "^2.0.3"
scikit-learn = "^1.3.0"
opencv-python = "^4.8.0.74"
pillow = "^10.0.0"
tqdm = "^4.65.0"
albumentations = "^1.3.1"
timm = "^0.9.5"
tensorboardx = "^2.6.2"
pydantic = "^2.3.0"
pilgram = "^1.2.1"
matplotlib = "^3.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
